refactor(session): 移除 session 文件夹及下面的全部文件， 回滚为统一引用 session 模块
- 删除了整个 session 文件夹及相关类型定义
- 统一所有模块中对 session 相关类型的导入，改为从 session.ts 模块直接导入
