+++ src/session.ts
@@ -1,3 +1,224 @@
+import type { McpServerConfig, PermissionScope, PermissionSettings } from "./settings";
+import type { AskPermissionRequest, MessageToolPermission, UserToolPermission } from "./common/permissions";
+import type { CreateOpenAIClient } from "./tools/executor";
+
+export type SessionStatus =
+  | "failed"
+  | "pending"
+  | "processing"
+  | "waiting_for_user"
+  | "completed"
+  | "interrupted"
+  | "ask_permission"
+  | "permission_denied";
+
+export type ModelUsage = {
+  prompt_tokens: number;
+  completion_tokens: number;
+  total_tokens: number;
+  completion_tokens_details?: Record<string, unknown>;
+  prompt_tokens_details?: Record<string, unknown>;
+  prompt_cache_hit_tokens?: number;
+  prompt_cache_miss_tokens?: number;
+  total_reqs?: number;
+};
+
+export type SessionProcessEntry = {
+  startTime: string;
+  command: string;
+  timeoutMs?: number;
+  deadlineAt?: string;
+  timedOut?: boolean;
+};
+
+export type BashTimeoutAdjustment = {
+  processId: string;
+  timeoutMs: number;
+  deadlineAt: string;
+  timedOut: boolean;
+};
+
+export type SessionEntry = {
+  id: string;
+  summary: string | null;
+  assistantReply: string | null;
+  assistantThinking: string | null;
+  assistantRefusal: string | null;
+  toolCalls: unknown[] | null;
+  status: SessionStatus;
+  failReason: string | null;
+  usage: ModelUsage | null;
+  usagePerModel: Record<string, ModelUsage> | null;
+  activeTokens: number;
+  createTime: string;
+  updateTime: string;
+  processes: Map<string, SessionProcessEntry> | null;
+  askPermissions?: AskPermissionRequest[];
+};
+
+export type SessionsIndex = {
+  version: 1;
+  entries: SessionEntry[];
+  originalPath: string;
+};
+
+export type SessionMessageRole = "system" | "user" | "assistant" | "tool";
+
+export type MessageMeta = {
+  function?: unknown;
+  paramsMd?: string;
+  resultMd?: string;
+  asThinking?: boolean;
+  isSummary?: boolean;
+  isModelChange?: boolean;
+  skill?: SkillInfo;
+  permissions?: MessageToolPermission[];
+  userPrompt?: UserPromptContent;
+};
+
+export type SessionMessage = {
+  id: string;
+  sessionId: string;
+  role: SessionMessageRole;
+  content: string | null;
+  contentParams: unknown | null;
+  messageParams: unknown | null;
+  compacted: boolean;
+  visible: boolean;
+  createTime: string;
+  updateTime: string;
+  meta?: MessageMeta;
+  html?: string;
+  checkpointHash?: string;
+};
+
+export type UndoTarget = {
+  message: SessionMessage;
+  index: number;
+  canRestoreCode: boolean;
+};
+
+export type UserPromptContent = {
+  text?: string;
+  imageUrls?: string[];
+  skills?: SkillInfo[];
+  permissions?: UserToolPermission[];
+  alwaysAllows?: PermissionScope[];
+};
+
+export type SkillInfo = {
+  name: string;
+  path: string;
+  description: string;
+  isLoaded?: boolean;
+};
+
+export type SessionManagerOptions = {
+  projectRoot: string;
+  createOpenAIClient: CreateOpenAIClient;
+  getResolvedSettings: () => {
+    model: string;
+    webSearchTool?: string;
+    mcpServers?: Record<string, McpServerConfig>;
+    permissions?: Required<PermissionSettings>;
+  };
+  renderMarkdown: (text: string) => string;
+  onAssistantMessage: (message: SessionMessage, shouldConnect: boolean) => void;
+  onSessionEntryUpdated?: (entry: SessionEntry) => void;
+  onLlmStreamProgress?: (progress: LlmStreamProgress) => void;
+  onMcpStatusChanged?: () => void;
+  onProcessStdout?: (pid: number, chunk: string) => void;
+};
+
+export type LlmStreamProgress = {
+  requestId: string;
+  sessionId?: string;
+  startedAt: string;
+  estimatedTokens: number;
+  formattedTokens: string;
+  phase: "start" | "update" | "end";
+};
+import { DEEPSEEK_V4_MODELS } from "./common/model-capabilities";
+
+const DEFAULT_COMPACT_PROMPT_TOKEN_THRESHOLD = 128 * 1024;
+const DEEPSEEK_V4_COMPACT_PROMPT_TOKEN_THRESHOLD = 512 * 1024;
+
+export function getCompactPromptTokenThreshold(model: string): number {
+  return DEEPSEEK_V4_MODELS.has(model)
+    ? DEEPSEEK_V4_COMPACT_PROMPT_TOKEN_THRESHOLD
+    : DEFAULT_COMPACT_PROMPT_TOKEN_THRESHOLD;
+}
+
+export function isUsageRecord(value: unknown): value is Record<string, unknown> {
+  return value !== null && typeof value === "object" && !Array.isArray(value);
+}
+
+export function getTotalTokens(usage: ModelUsage | null | undefined): number {
+  if (!isUsageRecord(usage)) {
+    return 0;
+  }
+  const totalTokens = (usage as Record<string, unknown>).total_tokens;
+  return typeof totalTokens === "number" ? totalTokens : 0;
+}
+
+export function summarizeCompletionOptions(options?: Record<string, unknown>): Record<string, unknown> | undefined {
+  if (!options) {
+    return undefined;
+  }
+  return {
+    ...options,
+    signal: options.signal instanceof AbortSignal ? { aborted: options.signal.aborted } : options.signal,
+  };
+}
+
+export function addUsageValue(current: unknown, next: unknown): unknown {
+  if (typeof next === "number") {
+    return (typeof current === "number" ? current : 0) + next;
+  }
+
+  if (isUsageRecord(next)) {
+    const currentRecord = isUsageRecord(current) ? current : {};
+    const result: Record<string, unknown> = { ...currentRecord };
+    for (const [key, value] of Object.entries(next)) {
+      result[key] = addUsageValue(currentRecord[key], value);
+    }
+    return result;
+  }
+
+  return next;
+}
+
+export function accumulateUsage(current: ModelUsage | null, next: unknown | null | undefined): ModelUsage | null {
+  if (next == null) {
+    return current ?? null;
+  }
+  return addUsageValue(current, next) as ModelUsage;
+}
+
+export function usageWithRequestCount(usage: ModelUsage): ModelUsage {
+  const totalReqs = typeof usage.total_reqs === "number" ? usage.total_reqs + 1 : 1;
+  return {
+    ...usage,
+    total_reqs: totalReqs,
+  };
+}
+
+export function accumulateUsagePerModel(
+  current: Record<string, ModelUsage> | null | undefined,
+  model: string,
+  next: ModelUsage | null | undefined
+): Record<string, ModelUsage> | null {
+  if (next == null) {
+    return current ?? null;
+  }
+
+  const usagePerModel = { ...(current ?? {}) };
+  const modelName = model.trim() || "unknown";
+  usagePerModel[modelName] = accumulateUsage(usagePerModel[modelName] ?? null, usageWithRequestCount(next))!;
+  return usagePerModel;
+}
+
+export { getExtensionRoot } from "./prompt";
 import * as fs from "fs";
 import * as path from "path";
 import * as os from "os";
@@ -5,68 +226,41 @@ import * as crypto from "crypto";
 import matter from "gray-matter";
 import ejs from "ejs";
 import type { ChatCompletionContentPart, ChatCompletionMessageParam } from "openai/resources/chat/completions";
-import { launchNotifyScript } from "../common/notify";
-import { buildThinkingRequestOptions } from "../common/openai-thinking";
-import { supportsMultimodal } from "../common/model-capabilities";
+import { launchNotifyScript } from "./common/notify";
+import { buildThinkingRequestOptions } from "./common/openai-thinking";
+import { supportsMultimodal } from "./common/model-capabilities";
 import {
   getCompactPrompt,
   getDefaultSkillPrompt,
+  getExtensionRoot,
   getRuntimeContext,
   getSystemPrompt,
   getTools,
   type ToolDefinition,
-} from "../prompt";
+} from "./prompt";
 import {
-  type CreateOpenAIClient,
   type ProcessTimeoutControl,
   type ProcessTimeoutInfo,
   type ToolCallExecution,
   type ToolExecutionHooks,
   ToolExecutor,
-} from "../tools/executor";
-import { McpManager } from "../mcp/mcp-manager";
-import type { McpServerConfig, PermissionSettings } from "../settings";
-import { logApiError } from "../common/logging/error-logger";
-import { logOpenAIChatCompletionDebug, normalizeDebugError } from "../common/logging/debug-logger";
-import { killProcessTree } from "../common/system/process-tree";
-import { GitFileHistory } from "../common/runtime/file-history";
-import { getSnippet } from "../common/runtime/state";
+} from "./tools/executor";
+import { McpManager } from "./mcp/mcp-manager";
+
+import { logApiError } from "./common/logging/error-logger";
+import { logOpenAIChatCompletionDebug, normalizeDebugError } from "./common/logging/debug-logger";
+import { killProcessTree } from "./common/system/process-tree";
+import { GitFileHistory } from "./common/runtime/file-history";
+import { getSnippet } from "./common/runtime/state";
 import {
   appendProjectPermissionAllows,
   buildPermissionToolExecution,
   computeToolCallPermissions,
   hasUserPermissionReplies,
-  type MessageToolPermission,
   normalizeAskPermissions,
   parseToolCallForPermissions,
   type PermissionToolCall,
-  type UserToolPermission,
-} from "../common/permissions";
-
-import {
-  accumulateUsage,
-  accumulateUsagePerModel,
-  getCompactPromptTokenThreshold,
-  getExtensionRoot,
-  getTotalTokens,
-  isUsageRecord,
-  summarizeCompletionOptions,
-} from "./utils";
-import {
-  type BashTimeoutAdjustment,
-  type LlmStreamProgress,
-  type MessageMeta,
-  type ModelUsage,
-  type SessionEntry,
-  type SessionManagerOptions,
-  type SessionMessage,
-  type SessionProcessEntry,
-  type SessionsIndex,
-  type SessionStatus,
-  type SkillInfo,
-  type UndoTarget,
-  type UserPromptContent,
-} from "./types";
+} from "./common/permissions";
 
 const MAX_SESSION_ENTRIES = 50;
 const DEFAULT_NEW_PROMPT_API_URL = "https://deepcode.vegamo.cn/api/plugin/new";
