+++ src/ui/components/MessageView/types.ts
@@ -1,4 +1,4 @@
-import type { SessionMessage } from "../../../session/types";
+import type { SessionMessage } from "../../../session";
 
 export type MessageViewProps = {
   message: SessionMessage;
