+++ src/ui/core/loading-text.ts
@@ -1,4 +1,4 @@
-import type { LlmStreamProgress, SessionEntry } from "../../session/types";
+import type { LlmStreamProgress, SessionEntry } from "../../session";
 
 type RunningProcesses = SessionEntry["processes"];
