+++ src/ui/utils/index.ts
@@ -3,7 +3,7 @@ import { renderMessageToStdout } from "../components/MessageView/utils";
 import type { RawMode } from "../contexts";
 import type { PromptDraft } from "../views/PromptInput";
 import type { ModelConfigSelection } from "../../settings";
-import type { SessionEntry, SessionMessage } from "../../session/types";
+import type { SessionEntry, SessionMessage } from "../../session";
 import type { SessionManager } from "../../session";
 
 /**
