+++ src/ui/views/ProcessStdoutView.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { Box, Text } from "ink";
 import { BASH_TIMEOUT_DECREMENT_MS, BASH_TIMEOUT_INCREMENT_MS } from "../../common/system/bash-timeout";
-import type { BashTimeoutAdjustment, SessionEntry, SessionProcessEntry } from "../../session/types";
+import type { BashTimeoutAdjustment, SessionEntry, SessionProcessEntry } from "../../session";
 import { useTerminalInput } from "../hooks";
 
 type RunningProcesses = SessionEntry["processes"];
