+++ src/ui/views/SessionList.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useMemo, useCallback } from "react";
 import { Box, Text, useInput, useWindowSize } from "ink";
-import type { SessionEntry, SessionStatus } from "../../session/types";
+import type { SessionEntry, SessionStatus } from "../../session";
 import { truncate } from "../components/MessageView/utils";
 
 type Props = {
