+++ src/ui/views/UndoSelector.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo, useState } from "react";
 import { Box, Text, useInput, useWindowSize } from "ink";
-import type { UndoTarget } from "../../session/types";
+import type { UndoTarget } from "../../session";
 
 export type UndoRestoreMode = "code-and-conversation" | "conversation";
