+++ src/ui/views/WelcomeScreen.tsx
@@ -2,7 +2,7 @@ import React, { useMemo, useState } from "react";
 import { Box, Text } from "ink";
 import * as os from "node:os";
 import path from "node:path";
-import type { SkillInfo } from "../../session/types";
+import type { SkillInfo } from "../../session";
 import type { ResolvedDeepcodingSettings } from "../../settings";
 import { buildSlashCommands, BUILTIN_SLASH_COMMANDS, formatSlashCommandDescription } from "../core/slash-commands";
 import { ThemedGradient } from "./ThemedGradient";
